feat(navbar): added a navbar to the page (#2)
Added a navbar with my name as a logo and some links to sections that will be included in the main section of the page. The navbar is responsive to window width, where a burger menu appears if the width reaches a size of 45em. The burger menu can be clicked to bring a dropdown of all the links. The dropdown is automatically removed when the screen width is increased again, such that the links appear in the navbar again.