feat(about-section): added about section to page (#5)
Added the about section to the website which includes 4 flip cards. Flip cards have two sides, the first side shown is just the heading representing the content within that card. The flip card can be clicked and flipped to its other side to show the content it has inside. There is a flip card with the following headings:

- Skills
- Achievements
- Interests
- Hobbies