fix(fix-compatibility): fixed design issues incompatible with Safari #6
Replaced all the gap properties in all flexbox divs with manually set margins that simulate its effect. This is because gap is not supported for versions of Safari that are earlier than 14.1. Also changed the burger menu to be a regular div rather than a button since it seemed to not be visible as a button on Safari 13.0.4, but also remains consistent with the fact that Flipcards are also not set as button tag.