feat(timeline): added a timeline section (#7)
Adds a timeline of my educational history of computer science in form of an animated timeline graph. There are timeline items that appear on the sides of the graph, with corresponding nodes put on the graph in the location of the item. The graph also fills in an orange colour as the user scrolls down the page. When a node in the graph is reached, it also lights up with an orange colour.