feat(projects): added projects section to page (#8)
The projects section consists of an image of a cloud which animates on hover. When clicked, it opens a project slideshow containing information about each of the personal projects I have done. The data for each project is stored in a json so that the slideshow is created dynamically and so adding new projects in the future is a trivial matter of updating the json and adding a new project image. The slideshow supports animation transitions in Chrome (but not in Safari).